requirements.txt
requests
beautifulsoup4
openai
pydantic
python-dotenv
responses
pytest
pytest-mock
pyyaml
litellm
